requirements.txt
asgiref==3.4.1
bleach==4.1.0
defusedxml==0.7.1
diff-match-patch==20200713
Django==3.1.7
django-import-export==2.6.1
django-summernote==0.8.20.0
et-xmlfile==1.1.0
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.0.9
packaging==21.3
Pillow==8.4.0
pyparsing==3.0.6
pytz==2021.3
PyYAML==6.0
six==1.16.0
sqlparse==0.4.2
tablib==3.1.0
typing_extensions==4.0.0
webencodings==0.5.1
xlrd==2.0.1
xlwt==1.3.0
